Camera matrix

+++ shaders/layouts.glsl
@@ -69,7 +69,7 @@ layout(set = S_RESTIR, binding = eIndexTemp, scalar)   buffer _IndexTempBuffer {
 layout(set = S_RESTIR, binding = eDebugUintImage)					uniform uimage2D			debugUintImage;
 layout(set = S_RESTIR, binding = eDebugImage)						uniform image2D			debugImage;
 layout(set = S_RESTIR, binding = eDebugUintBuffer)					buffer _DebugUintBuffer { uint debugUintBuffer[]; };
-layout(set = S_RESTIR, binding = eDebugFloatBuffer)					buffer _DebugFloatBuffer { uint debugFloatBuffer[]; };
+layout(set = S_RESTIR, binding = eDebugFloatBuffer)					buffer _DebugFloatBuffer { float debugFloatBuffer[]; };
 
 layout(buffer_reference, scalar) buffer Vertices { VertexAttributes v[]; };
 layout(buffer_reference, scalar) buffer Indices	 { uvec3 i[];            };
